pyproject.toml
[project]
name = "python-utils"
version = "0.3.0"
description = "A collection of Python utilities for PDF/image conversion, JSON processing, and AI agent initialization"
readme = "README.md"
authors = [
    { name = "Marc Fabian Mezger", email = "57255687+mfmezger@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.53.0",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic-ai>=1.26.0",
    "pymupdf>=1.26.6",
    "pymupdf4llm>=0.2.6",
    "tqdm>=4.67.1",
    "typer>=0.15.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
python-utils = "python_utils:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
